Created
YAMP is now on GitHub. The is what I worked on the last 2 days. The
parser is already quite good and the included functions and objects are
working quite well. What is lacking right now: Speed (could be faster)
and assignment (the operator is there, but I still have to think about
two or three small things).