Updated portable library and replaced Math.Truncate with (int)
Math.Truncate is not supported in the Portable subset. Casting to int
yield equivalent results and is faster than Math.Truncate. Within the
scope of the Gamma.LogGamma_Stirling method, the cast/truncated value
should always be within the 32 -bit integer range, so casting to int
should be safe from a numerical overflow perspective.