1.2.0 FINAL
Repositioned some functions in the physics library, improved the event
model, added some functions (with events) to prompt for user input and
to cast some string to a scalar (more maybe to follow). Improved the
documentation and some numeric algorithms. Improved the parser and added
some functions like the Faddeeva function. Improved the existing plot
values and the loading process.