YAMP 1.2.2 second
Fixed a bug in the number expression concerning exponential notation.
Added more Wiki-Links, improved the visualization function and changed
some return types to "void". Fixed a bug in the laplace distribution and
improved the "variable change" event (did not catch expanding matrices).
Improved the subplot to use the ChangeLastPlotTo function and created
the new setdef function for templating plots.